+++ src/pages/Admin.jsx
@@ -1,7 +1,10 @@
 import React from 'react'
 import { Outlet } from 'react-router-dom';
 import { IoIosNotificationsOutline,IoMdMenu  } from "react-icons/io";
-import { IoPerson } from "react-icons/io5"
+import { IoPerson } from "react-icons/io5"  
+import { adminUrls } from '../Routes/UrlRoutes';
+import { NavLink } from 'react-router-dom';
+import '../components/NavBar/NavBar.css'
 
 const Admin = () => {
   return (
@@ -15,9 +18,9 @@ const Admin = () => {
 
         <div className="flex main-contain ">
             <div className="sidebar w-[15%] h-[93vh] flex flex-col bg-gray-300 overflow-hidden justify-center items-center  ">
-            <a href="/">Orders</a>
-            <a href="/">Add Users</a>
-            <a href="/">click</a>
+            {adminUrls.map((route, index)=>{
+                return <NavLink to={route.path} key={index} className={({isActive})=>(isActive?'active-link':'link')}>{route.title}</NavLink>
+            })}
             </div>
 
             <div className="mx-8 my-4 h-[88vh] content w-[80%] ">
